+++ packages/common/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Add soroban to network family (#1919)
+- Add stellar to network family (#1919)
 
 ## [2.4.0] - 2023-07-31
 ### Fixed